+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add: Query parameter for all collection endpoints. (skip and limit your responses)
 - Add: Ability to track rate limit headers.
 - Fix: Return HTTP Code of no content methods.
+- Fix: Assets: Do not send `url` and `details` if `uploadUrl` or `uploadFrom` is changed.
 - Polishing: Update dependencies
   - commonsio 2.5 (was 2.4)
   - junit: 4.12 (was 4.11)